Remove testing symbols from lib/service (#57262)
`func (*TeleportProcess) GetIdentityForTesting` was the only place
consuming test symbols outside of tests. It existed to allow
`integration.TestInstanceCertReissue` to assert the identity after
certificate reissuance. To avoid this, TestInstanceCertReissue
was moved from integration into lib/service so that it could make
use of the private `(*TeleportProcess) getIdentity` and allow the
`GetIdentityForTesting` to be removed.

Updates #51023.
