+++ lib/service/testdata/auth/host_uuid
